LAB8/requirements.txt
flask
psycopg2-binary
flask_swagger_ui
requests
